app/requirements.txt
Django>=4.2
djangorestframework
mysqlclient
python-dotenv
djangorestframework-simplejwt
requests
drf-spectacular>=0.27.0
django-cors-headers
